requirements.txt
flask
pandas
smartapi-python
pyotp
requests
TA-Lib
logzero
websocket